requirements.txt
h5py==2.10.0
matplotlib==3.1.1
numpy==1.17.2
pandas==0.25.1
pymia==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
SimpleITK==1.2.2
tensorboardX==1.9
torch==1.2.0
Pillow==6.2.0